+++ test-parts/test-pathfinding-input.js
@@ -220,8 +220,8 @@ function handlePossibleRoutes()
 		
 		it("Impossible Route", function()
 		{
-			//var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "G", emptyCriteria, emptyCriteria);
-			//console.log(resultValue);
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "G", emptyCriteria, emptyCriteria);
+			console.log(resultValue);
 			expect(true).to.be.true;
 		});